Revert "Add view history and validation part to user form"
