+++ src/components/SideBar.jsx
@@ -79,7 +79,6 @@ const SideBar = () => {
       </Collapse>
 
       {/* User */}
-      <a href="/user">
       <ListItem
         button
         className="rounded-lg hover:bg-blue-100 focus:bg-blue-400"
@@ -87,7 +86,6 @@ const SideBar = () => {
         <PersonIcon></PersonIcon>
         <ListItemText primary="User" className="pl-4 pr-4" />
       </ListItem>
-      </a>
 
       {/* Request & Reservation with dropdown and sub-parts */}
       <ListItem
@@ -128,12 +126,9 @@ const SideBar = () => {
       </ListItem>
       <Collapse in={openReport} timeout="auto" unmountOnExit>
         <List component="div" disablePadding>
-          <a href="/history">
           <ListItem button className="pl-8 rounded-lg">
             <ListItemText primary="View History" />
-        
           </ListItem>
-          </a>
           <ListItem button className="pl-8 rounded-lg">
             <ListItemText primary="Inventory Summary" />
           </ListItem>
