+++ src/pages/Home.jsx
@@ -1,84 +1,45 @@
-import React from "react";
-import NavBar from "../components/NavBar";
-import SideBar from "../components/SideBar";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-//import Dashboard from './Dashboard'
-import Adjustment from "./Adjustment/Adjustment";
-import NewAdjustment from "./Adjustment/NewAdjustment";
-import User from "./User/User";
-import CreateUser from "./User/CreateUser";
-import Userupdate from "./User/Edit_user";
-import EditUser from "../components/User_Components/UserEditForm";
-import { Grid } from "@mui/material";
-import AdjustmentDocument from "./Adjustment/AdjustmentDocument";
-import RequestList from "./InventoryRequest/InventoryRequestList";
-import ItemDataGrid from "./InventoryItem/InventoryItems";
-import AddItemForm from "./InventoryItem/NewItem";
-import ViewItemDetails from "./InventoryItem/ViewDetails";
-import ViewHistory from "./User/History";
-import Changepassword from "./User/ChangePassword";
-import TicketForm from "../components/Ticket_Create";
-import { BrowserRouter as Router, Link } from "react-router-dom";
+import React from 'react'
+import NavBar from '../components/NavBar'
+import SideBar from '../components/SideBar'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import AdminDashboard from './AdminDashboard'
+import NewAdjustment from './Adjustment/NewAdjustment'
+import User from './User/User'
+import CreateUser from './User/CreateUser'
+
 import EditUser from '../components/UserEditForm'
+import { Grid } from '@mui/material'
+import AdjustmentDocument from './Adjustment/AdjustmentDocument'
 import RequestList from './InventoryRequest/RequestsList'
+import ItemDataGrid from './InventoryItem/InventoryItems'
+import AddItemForm from './InventoryItem/NewItem'
+import ViewItemDetails from './InventoryItem/ViewDetails'
+
 import RequestDocumentAccept from './InventoryRequest/RequestDocumentAccept'
 import CreateNewRequest from './InventoryRequest/CreateNewRequest'
 import EditRequest from './InventoryRequest/EditRequest'
+
 import EditAdjustment from './Adjustment/EditAdjustment'
 import StockInList from './StockIn/StockInList'
 import StockOutList from './StockOut/StockOutList'
 import AdjustmentList from './Adjustment/AdjustmentList'
 
+
+
 const Home = () => {
   return (
     <div>
       <div>
-      <Router>
-        <NavBar />
-      </Router>
+        <NavBar/>
       </div>
       <Grid container className="flex">
-        <Grid
-          item
-          sm={2.5}
-          className="box-border hidden h-screen md:block w-96"
-        >
+        <Grid item sm={2.5} className='box-border hidden h-screen md:block w-96'>
           <SideBar></SideBar>
         </Grid>
 
-        <Grid
-          item
-          sm={9.5}
-          style={{ backgroundColor: "#eeeeee" }}
-          className="w-screen p-10"
-        >
-          {/* Enter components here, that you want to insert. */}
+        <Grid item sm={9.5} style={{ backgroundColor: '#eeeeee' }} className='w-screen p-10'>
           <BrowserRouter>
             <Routes>
-
-              {/* <Route path='/' element={<Dashboard name="Dashboard"/>}></Route> */}
-              <Route path="/adjustment" element={<Adjustment />}></Route>
-              
-              <Route path="/user" element={<User />}></Route>
-              <Route path="/newUser" element={<CreateUser />} />
-              <Route path="/user/users/:ID" element={<EditUser />} />
-              <Route path="/newadjustment" element={<NewAdjustment />}></Route>
-              <Route
-                path="/adjustment/adj1"
-                element={<AdjustmentDocument />}
-              ></Route>
-              <Route path="/inventoryRequest" element={<RequestList />}></Route>
-              <Route path="/item" element={<ItemDataGrid />}></Route>
-              <Route path="/item/add-item" element={<AddItemForm />}></Route>
-              <Route
-                path="/item/edit-item/:ID"
-                element={<ViewItemDetails />}
-              ></Route>
-              <Route path="/history" element={<ViewHistory />}></Route>
-              <Route path="/changepassword" element={<Changepassword />}></Route>
-              <Route path="/newTicket" element={<TicketForm/>}></Route>
-
               {/* Enter components here, that you want to insert. */}
               {/* Dashboard routing */}
               <Route path='/' element={<AdminDashboard/>}></Route>
@@ -116,13 +77,12 @@ const Home = () => {
               {/* report routing */}
               {/* <Route path='/report' element={<StockLineChart/>}></Route>
                */}
-
             </Routes>
           </BrowserRouter>
         </Grid>
       </Grid>
     </div>
-  );
-};
+  )
+}
 
-export default Home;
+export default Home
