+++ src/pages/User/CreateUser.jsx
@@ -1,9 +1,5 @@
 import React from "react";
-import UserForm from "../../components/User_Components/UserForm";
-import NavBar from "../../components/NavBar";
-import SideBar from "../../components/SideBar";
 import UserForm from "../../components/UserForm";
-
 
 
 const CreateUser=()=>{
