+++ src/pages/User/User.jsx
@@ -1,9 +1,4 @@
 import React from 'react'
-import { useNavigate } from "react-router-dom";
-import NavBar from '../../components/NavBar'
-import SideBar from '../../components/SideBar'
-import {Button} from "@mui/material";
-import UserTable from '../../components/User_Components/UserTable';
 import UserTable from '../../components/UserTable';
 import { Box } from '@mui/material'
